18 studies of correlation coefficients reported by Craft et al. (2003)

This dataset includes 18 studies of correlation coefficients reported by Craft, Magyar, Becker, and Feltz (2003).

data(Craft2003)

Details

The main purpose of Craft and colleagues (2003) meta-analysis was to examine the interrelationships between athletic performance and three subscales, cognitive anxiety, somatic anxiety, and self-concept, of the Competitive State Anxiety Inventory (CSAI 2; CITATION). In this meta-analysis, the correlation coefficient was the primary effect size measure. For the purpose of demonstration, I use a subset of the data, i.e., six correlation coefficients among cognitive anxiety, somatic anxiety, self-concept, and sport performance in athletes.

ID	ID for each study included

N	sample size from each study included

gender	gender

p_male	percentage of male

C1	Correlation coefficient between cognitive anxiety and somatic anxiety

C2	Correlation coefficient between cognitive anxiety and self concept

C3	Correlation coefficient between cognitive anxiety and athletic performance

C4	Correlation coefficient between somatic anxiety and self concept

C5	Correlation coefficient between somatic anxiety and athletic performance

C6	Correlation coefficient between self concept and athletic performance

Source

Ahn, S., Lu, M., Lefevor, G.T., Fedewa, A. & Celimli, S. (2016). Application of meta-analysis in sport and exercise science. In N. Ntoumanis, & N. Myers (Eds.), An Introduction to Intermediate and Advanced Statistical Analyses for Sport and Exercise Scientists (pp.233-253). Hoboken, NJ: John Wiley and Sons, Ltd.

Craft, L. L., Magyar, T. M., Becker, B. J., & Feltz, D. L. (2003). The relationship between the competitive state anxiety inventory-2 and sport performance: a meta-analysis. Journal of Sport and Exercise Psychology, 25(1), 44-65.

Examples

  data(Craft2003)