Multivariate Effect Sizes of Different Types from the Geeganage et al. Meta-Analysis

This dataset includes 17 studies of multivariate effect sizes with four different outcomes reported by Geeganage and Bath (2010).

data(Geeganage2010)

Details

In a meta-analysis, Geeganage and Bath (2010) studied whether blood pressure (BP) should be actively altered during the acute phase of stroke, and assessed the effect of multiple vasoactive drugs on BP in acute stroke. Selection criteria included: Randomized trials of interventions that would be expected, on pharmacological grounds, to alter BP in patients within one week of the onset of acute stroke. There were four outcomes: systolic blood pressure (SBP, in mHg), diastolic blood pressure (DBP, in mHg), death (D), and death or disability (DD).

ID:	ID for each study included

ft_D	Number of early death within 1 month (D) in "1 Drug" Group

fc_D	Number of D in "control" Group

nt_D	Number of people in "1 Drug" Group reporting D status

nc_D	Number of people in "control" Group reporting D status

OR_D	Odds ratio of D for "1 Drug" versus "control" group

ft_DD	Number of early death or deterioration within 1 month (DD) in "1 Drug" Group

fc_DD	Number of early DD in "control" Group

nt_DD	Number of people in "1 Drug" Group reporting DD status

nc_DD	Number of people in "control" Group reporting DD status

OR_DD	Odds ratio of DD for "1 Drug" versus "control" group

nt_SBP	Number of people in "1 Drug" Group reporting Systolic blood pressure (SBP) status

nc_SBP	Number of people in "control" Group reporting SBP status

MD_SBP	Mean difference of SBP for "1 Drug" versus "control" group

sdt_SBP	Standard deviation of SBP in "1 Drug" Group

sdc_SBP	Standard deviation of SBP in "control" Group

nt_DBP	Number of people in "1 Drug" Group reporting Diastolic blood pressure (DBP) status

nc_DBP	Number of people in "control" Group reporting DBP status

MD_DBP	Mean difference of DBP for "1 Drug" versus "control" group

sdt_DBP	Standard deviation of DBP in "1 Drug" Group

sdc_DBP	Standard deviation of DBP in "control" Group

SMD_SBP	Standardized mean difference of SBP for "1 Drug" versus "control" group

SMD_DBP	Standardized mean difference of DBP for "1 Drug" versus "control" group

Source

Geeganage, C., & Bath, P. M. (2010). Vasoactive drugs for acute stroke. Cochrane Database of Systematic Reviews 2010(7).

Examples

  data(Geeganage2010)