Covariance between mean difference and log risk ratio

Compute covariance between mean difference and log risk ratio, when effect sizes are different.

md_lgrr(r, n1c, n2c, n1t, n2t,
        n12c = min(n1c, n2c), n12t = min(n1t, n2t),
        s2c, s2t, f2c, f2t, sd1c, sd1t)

Arguments

r
Correlation coefficient of the two outcomes.

n1c
Number of participants reporting outcome 1 in control group.

n2c
Number of participants reporting outcome 2 in control group.

n1t
Number of participants reporting outcome 1 in treatment group.

n2t
Number of participants reporting outcome 2 in treatment group.

n12c
Number of participants reporting both outcome 1 and outcome 2 in control group. By default, it is equal to the smaller number between n1c and n2c.

n12t
Number defined in a similar way as n12c for treatment group.

s2c
Number of participants with event for outcome 2 (dichotomous) in control group.

s2t
Defined in a similar way as s2c for treatment group.

f2c
Number of participants without event for outcome 2 (dichotomous) in control group.

f2t
Defined in a similar way as f2c for treatment group.

sd1c
Sample standard deviation of outcome 1.

sd1t
Defined in a similar way as sd1c for treatment group.

Author

Min Lu

Value

Return the computed covariance.

References

Ahn, S., Lu, M., Lefevor, G.T., Fedewa, A. & Celimli, S. (2016). Application of meta-analysis in sport and exercise science. In N. Ntoumanis, & N. Myers (Eds.), An Introduction to Intermediate and Advanced Statistical Analyses for Sport and Exercise Scientists (pp.233-253). Hoboken, NJ: John Wiley and Sons, Ltd.

Wei, Y., & Higgins, J. (2013). Estimating within study covariances in multivariate meta-analysis with multiple outcomes. Statistics in Medicine, 32(7), 119-1205.

Examples

## simple example
md_lgrr(r = 0.71, n1c = 34, n2c = 35, n1t = 25, n2t = 32,
        s2c = 5, s2t = 8, f2c = 30, f2t = 24, sd1t = 0.4, sd1c = 8)
## calculate covariances for variable SBP and DD in Geeganage2010 data
attach(Geeganage2010)
SBP_DD <- unlist(lapply(1:nrow(Geeganage2010), function(i){md_lgrr(r = 0.71,
                n1c = nc_SBP[i], n2c = nc_DD[i], n1t = nt_SBP[i], n2t = nt_DD[i],
                sd1t = sdt_SBP[i], s2t = st_DD[i], sd1c = sdc_SBP[i], s2c = sc_DD[i],
                f2c = nc_DD[i] - sc_DD[i], f2t = nt_DD[i] - st_DD[i])}))
SBP_DD