Computing Covariance between Mean Difference and Standardized Mean Difference

The function lgor_rd computes covariance between mean difference and standardized mean difference. See mix.vcov for effect sizes of the same or different types.

md_smd(smd, r, n1c, n2c, n1t, n2t,
       n12c = min(n1c, n2c), n12t = min(n1t, n2t),
       sd1t, sd2t, sd1c, sd2c)

Arguments

smd
Standardized mean difference for outcome 2.

r
Correlation coefficient of the two outcomes.

n1c
Number of participants reporting outcome 1 in the control group.

n2c
Number of participants reporting outcome 2 in the control group.

n1t
Number of participants reporting outcome 1 in the treatment group.

n2t
Number of participants reporting outcome 2 in the treatment group.

n12c
Number of participants reporting both outcome 1 and outcome 2 in the control group. By default, it is equal to the smaller number between n1c and n2c.

n12t
Number defined in a similar way as n12c for treatment group.

sd1t
Sample standard deviation of outcome 1 for the treatment group.

sd2t
Sample standard deviation of outcome 2 for the treatment group.

sd1c
Defined in a similar way as sd1t for the control group.

sd2c
Defined in a similar way as sd2t for the control group.

Author

Min Lu

Value

g	Computed Hedge's g from the input argument smd for outcome 2.

v	Computed covariance.

References

Ahn, S., Lu, M., Lefevor, G.T., Fedewa, A. & Celimli, S. (2016). Application of meta-analysis in sport and exercise science. In N. Ntoumanis, & N. Myers (Eds.), An Introduction to Intermediate and Advanced Statistical Analyses for Sport and Exercise Scientists (pp.233-253). Hoboken, NJ: John Wiley and Sons, Ltd.

Wei, Y., & Higgins, J. (2013). Estimating within study covariances in multivariate meta-analysis with multiple outcomes. Statistics in Medicine, 32(7), 119-1205.

Examples

## a simple example
md_smd(smd = 1, r = 0.71, n1c = 34, n2c = 35, n1t = 25, n2t = 32,
       sd1t = 0.6, sd2t = 0.4, sd1c = 8, sd2c = 0.9)

#> $g
#> [1] 1.013393
#> 
#> $v
#> [1] 0.02449068
#> 

## calculate covariances for variable SBP and DBP in Geeganage2010 data
attach(Geeganage2010)

#> The following objects are masked from Geeganage2010 (pos = 3):
#> 
#>     MD_DBP, MD_SBP, nc_D, nc_DBP, nc_DD, nc_SBP, nt_D, nt_DBP, nt_DD,
#>     nt_SBP, OR_D, OR_DD, sc_D, sc_DD, sdc_DBP, sdc_SBP, sdt_DBP,
#>     sdt_SBP, SMD_DBP, SMD_SBP, st_D, st_DD, studyID

#> The following objects are masked from Geeganage2010 (pos = 4):
#> 
#>     MD_DBP, MD_SBP, nc_D, nc_DBP, nc_DD, nc_SBP, nt_D, nt_DBP, nt_DD,
#>     nt_SBP, OR_D, OR_DD, sc_D, sc_DD, sdc_DBP, sdc_SBP, sdt_DBP,
#>     sdt_SBP, SMD_DBP, SMD_SBP, st_D, st_DD, studyID

#> The following objects are masked from Geeganage2010 (pos = 5):
#> 
#>     MD_DBP, MD_SBP, nc_D, nc_DBP, nc_DD, nc_SBP, nt_D, nt_DBP, nt_DD,
#>     nt_SBP, OR_D, OR_DD, sc_D, sc_DD, sdc_DBP, sdc_SBP, sdt_DBP,
#>     sdt_SBP, SMD_DBP, SMD_SBP, st_D, st_DD, studyID

#> The following objects are masked from Geeganage2010 (pos = 6):
#> 
#>     MD_DBP, MD_SBP, nc_D, nc_DBP, nc_DD, nc_SBP, nt_D, nt_DBP, nt_DD,
#>     nt_SBP, OR_D, OR_DD, sc_D, sc_DD, sdc_DBP, sdc_SBP, sdt_DBP,
#>     sdt_SBP, SMD_DBP, SMD_SBP, st_D, st_DD, studyID

#> The following objects are masked from Geeganage2010 (pos = 7):
#> 
#>     MD_DBP, MD_SBP, nc_D, nc_DBP, nc_DD, nc_SBP, nt_D, nt_DBP, nt_DD,
#>     nt_SBP, OR_D, OR_DD, sc_D, sc_DD, sdc_DBP, sdc_SBP, sdt_DBP,
#>     sdt_SBP, SMD_DBP, SMD_SBP, st_D, st_DD, studyID

#> The following objects are masked from Geeganage2010 (pos = 8):
#> 
#>     MD_DBP, MD_SBP, nc_D, nc_DBP, nc_DD, nc_SBP, nt_D, nt_DBP, nt_DD,
#>     nt_SBP, OR_D, OR_DD, sc_D, sc_DD, sdc_DBP, sdc_SBP, sdt_DBP,
#>     sdt_SBP, SMD_DBP, SMD_SBP, st_D, st_DD, studyID

SBP_DBP <- unlist(lapply(1:nrow(Geeganage2010), function(i){md_smd(smd = SMD_DBP, r = 0.71,
                n1c = nc_SBP[i], n2c = nc_DBP[i], n1t = nt_SBP[i], n2t = nt_DBP[i],
                sd1t = sdt_SBP[i], sd2t = sdt_DBP[i],
                sd1c = sdc_SBP[i], sd2c = sdc_SBP[i])$v}))
SBP_DBP

#>  [1] 1.98519067 2.34042056 0.41950881 0.42324831 0.25336382 3.88526087
#>  [7] 0.35189020 0.17785957 0.45956107 0.40763084 0.13144089 1.12143821
#> [13] 1.08637853 1.71396400 0.08663262 0.71054577 1.74388868

## the function mix.vcov() should be used for dataset