Computing Variance-Covariance Matrices for Correlation Coefficients

The function r.vcov computes variance-covariance matrix for multivariate meta-analysis when the effect size is measured by correlation coefficient.

r.vcov(n, corflat, zscore = FALSE, name = NULL, method = "average", na.impute = NA)

Arguments

n
A \(N\)-dimensional vector containing sample sizes from \(N\) studies.

corflat
A \(N \times p\) matrix or data frame storing \(p\) correlation coefficients from each of the \(N\) studies.

zscore
Whether the correlation coefficients in corflat are already transformed into Fisher's \(z\) scores.

name
A \(p\)-dimensional vector containing names for the \(p\) correlation coefficients.

method
Method "average" computes variance and covariances with mean correlation coefficients that are sample-size weighted from all the \(N\) studies (missing values are omitted); method "each" computes variance and covariances with each of the corresponding correlation coefficients.

na.impute
Missing values can be imputed by a numeric value, such as zero by setting na.impute = 0. With the default setting na.impute = NA, missing values are not imputed. If specifying na.impute = "average", missing values are imputed by mean correlation coefficients that are sample-size weighted from the complete records.

Author

Min Lu

Details

How to arrange correlation coefficients of each study from matrix to vector is in Cooper et al book page 385 to 386. Details for average method are in book of Cooper et al page 388. Let \(r_{ist}\) denote the sample correlation coefficient that describes the relationship between variables \(s\) and \(t\) in study \(i\). We can calculate its variance as \(var({{r}_{ist}})={{(1-\rho_{ist}^{2})}^{2}}/{{n}_{i}},\) and the covariance between two correlation coefficients is $$ cov({{r}_{ist}},{{r}_{iuv}})=[.5{{\rho }_{ist}}{{\rho }_{iuv}}({{\rho }}_{isu}^{2}+{{\rho }}_{isv}^{2}+{{\rho }}_{itu}^{2}+{{\rho }}_{itv}^{2})+{{\rho }_{isu}}{{\rho }_{itv}}+{{\rho }_{isv}}{{\rho }_{itu}} \nonumber \\ -({{\rho }_{ist}}{{\rho }_{isu}}{{\rho }_{isv}}+{{\rho }_{its}}{{\rho }_{itu}}{{\rho }_{itv}}+{{\rho }_{ius}}{{\rho }_{iut}}{{\rho }_{iuv}}+{{\rho }_{ivs}}{{\rho }_{ivt}}{{\rho }_{ivu}})]/{{n}_{i}}, $$ where \(\rho_{i..}\) represents the population value. In practice, \(\rho_{i..}\) can be substituted by the observed sample correlation or sample-size weighted mean correlation coefficients from all studies.

Value

r

A \(N \times p\) data frame that contains the input argument corflat with column names and imputed values according to the input argument na.impute. If the input argument zscore=TRUE, r is transformed from Fisher's \(z\) score in corflat.

list.rvcov

A \(N\)-dimensional list of \(p(p+1)/2 \times p(p+1)/2\) matrices of computed variance-covariance matrices.

matrix.rvcov

A \(N \times p(p+1)/2\) matrix whose rows are computed variance-covariance vectors.

ef

A \(N \times p\) data frame that contains Fisher's \(z\) transformed correlation coefficients from the input argument corflat with column names and imputed values according to the input argument na.impute.

list.vcov

A list in the same format of list.rvcov for Fisher's \(z\) transformed correlation coefficients.

matrix.vcov

A matrix matrix.rvcov for Fisher's \(z\) transformed correlation coefficients.

References

Ahn, S., Lu, M., Lefevor, G.T., Fedewa, A. & Celimli, S. (2016). Application of meta-analysis in sport and exercise science. In N. Ntoumanis, & N. Myers (Eds.), An Introduction to Intermediate and Advanced Statistical Analyses for Sport and Exercise Scientists (pp.233-253). Hoboken, NJ: John Wiley and Sons, Ltd.

Cooper, H., Hedges, L.V., & Valentine, J.C. (Eds.) (2009). The handbook of research synthesis and meta-analysis. New York: Russell Sage Foundation.

Olkin, I., & Ishii, G. (1976). Asymptotic distribution of functions of a correlation matrix. In S. Ikeda (Ed.), Essays in probability and statistics: A volume in honor of Professor Junjiro Ogawa (pp.5-51). Tokyo, Japan: Shinko Tsusho.

Examples