Computing Covariance between Standardized Mean Difference and Log Risk Ratio

The function smd_lgrr computes covariance between standardized mean difference and log risk ratio. See mix.vcov for effect sizes of the same or different types.

smd_lgrr(d, r, n1c, n2c, n1t, n2t,
         n12c = min(n1c, n2c), n12t = min(n1t, n2t),
         s2c, s2t, f2c, f2t, sd1c, sd1t)

Arguments

d
Standardized mean difference for outcome 1.

r
Correlation coefficient of the two outcomes.

n1c
Number of participants reporting outcome 1 in the control group.

n2c
Number of participants reporting outcome 2 in the control group.

n1t
Number of participants reporting outcome 1 in the treatment group.

n2t
Number of participants reporting outcome 2 in the treatment group.

n12c
Number of participants reporting both outcome 1 and outcome 2 in the control group. By default, it is equal to the smaller number between n1c and n2c.

n12t
Number defined in a similar way as n12c for the treatment group.

s2c
Number of participants with event for outcome 2 (dichotomous) in the control group.

s2t
Defined in a similar way as s2c for the treatment group.

f2c
Number of participants without event for outcome 2 (dichotomous) in the control group.

f2t
Defined in a similar way as f2c for the treatment group.

sd1c
Sample standard deviation of outcome 1 for the control group.

sd1t
Defined in a similar way as sd1c for the treatment group.

Author

Min Lu

Value

g	Computed Hedge's g from the input argument d for outcome 1.

lgrr	Computed log risk ratio for outcome 2.

v	Computed covariance.

References

Ahn, S., Lu, M., Lefevor, G.T., Fedewa, A. & Celimli, S. (2016). Application of meta-analysis in sport and exercise science. In N. Ntoumanis, & N. Myers (Eds.), An Introduction to Intermediate and Advanced Statistical Analyses for Sport and Exercise Scientists (pp.233-253). Hoboken, NJ: John Wiley and Sons, Ltd.

Wei, Y., & Higgins, J. (2013). Estimating within study covariances in multivariate meta-analysis with multiple outcomes. Statistics in Medicine, 32(7), 119-1205.

Examples

## simple example
smd_lgrr(d = 1, r = 0.3, n1c = 34, n2c = 35, n1t = 25, n2t = 32,
         s2c = 5, s2t = 8, f2c = 30, f2t = 24, sd1t = 0.4, sd1c = 8)

#> $g
#> [1] 1.013393
#> 
#> $lgrr
#> [1] 0.5596158
#> 
#> $v
#> [1] 0.02825625
#> 

## calculate covariances for variable SBP and DD in Geeganage2010 data
attach(Geeganage2010)

#> The following objects are masked from Geeganage2010 (pos = 3):
#> 
#>     MD_DBP, MD_SBP, nc_D, nc_DBP, nc_DD, nc_SBP, nt_D, nt_DBP, nt_DD,
#>     nt_SBP, OR_D, OR_DD, sc_D, sc_DD, sdc_DBP, sdc_SBP, sdt_DBP,
#>     sdt_SBP, SMD_DBP, SMD_SBP, st_D, st_DD, studyID

#> The following objects are masked from Geeganage2010 (pos = 7):
#> 
#>     MD_DBP, MD_SBP, nc_D, nc_DBP, nc_DD, nc_SBP, nt_D, nt_DBP, nt_DD,
#>     nt_SBP, OR_D, OR_DD, sc_D, sc_DD, sdc_DBP, sdc_SBP, sdt_DBP,
#>     sdt_SBP, SMD_DBP, SMD_SBP, st_D, st_DD, studyID

#> The following objects are masked from Geeganage2010 (pos = 8):
#> 
#>     MD_DBP, MD_SBP, nc_D, nc_DBP, nc_DD, nc_SBP, nt_D, nt_DBP, nt_DD,
#>     nt_SBP, OR_D, OR_DD, sc_D, sc_DD, sdc_DBP, sdc_SBP, sdt_DBP,
#>     sdt_SBP, SMD_DBP, SMD_SBP, st_D, st_DD, studyID

#> The following objects are masked from Geeganage2010 (pos = 9):
#> 
#>     MD_DBP, MD_SBP, nc_D, nc_DBP, nc_DD, nc_SBP, nt_D, nt_DBP, nt_DD,
#>     nt_SBP, OR_D, OR_DD, sc_D, sc_DD, sdc_DBP, sdc_SBP, sdt_DBP,
#>     sdt_SBP, SMD_DBP, SMD_SBP, st_D, st_DD, studyID

#> The following objects are masked from Geeganage2010 (pos = 10):
#> 
#>     MD_DBP, MD_SBP, nc_D, nc_DBP, nc_DD, nc_SBP, nt_D, nt_DBP, nt_DD,
#>     nt_SBP, OR_D, OR_DD, sc_D, sc_DD, sdc_DBP, sdc_SBP, sdt_DBP,
#>     sdt_SBP, SMD_DBP, SMD_SBP, st_D, st_DD, studyID

#> The following objects are masked from Geeganage2010 (pos = 11):
#> 
#>     MD_DBP, MD_SBP, nc_D, nc_DBP, nc_DD, nc_SBP, nt_D, nt_DBP, nt_DD,
#>     nt_SBP, OR_D, OR_DD, sc_D, sc_DD, sdc_DBP, sdc_SBP, sdt_DBP,
#>     sdt_SBP, SMD_DBP, SMD_SBP, st_D, st_DD, studyID

#> The following objects are masked from Geeganage2010 (pos = 12):
#> 
#>     MD_DBP, MD_SBP, nc_D, nc_DBP, nc_DD, nc_SBP, nt_D, nt_DBP, nt_DD,
#>     nt_SBP, OR_D, OR_DD, sc_D, sc_DD, sdc_DBP, sdc_SBP, sdt_DBP,
#>     sdt_SBP, SMD_DBP, SMD_SBP, st_D, st_DD, studyID

#> The following objects are masked from Geeganage2010 (pos = 13):
#> 
#>     MD_DBP, MD_SBP, nc_D, nc_DBP, nc_DD, nc_SBP, nt_D, nt_DBP, nt_DD,
#>     nt_SBP, OR_D, OR_DD, sc_D, sc_DD, sdc_DBP, sdc_SBP, sdt_DBP,
#>     sdt_SBP, SMD_DBP, SMD_SBP, st_D, st_DD, studyID

SBP_DD <- unlist(lapply(1:nrow(Geeganage2010), function(i){smd_lgrr(d = SMD_SBP, r = 0.3,
                 n1c = nc_SBP[i], n2c = nc_DD[i], n1t = nt_SBP[i], n2t = nt_DD[i],
                 sd1t = sdt_SBP[i], s2t = st_DD[i], sd1c = sdc_SBP[i], s2c = sc_DD[i],
                 f2c = nc_DD[i] - sc_DD[i], f2t = nt_DD[i] - st_DD[i])$v}))
SBP_DD

#>  [1] 0.044506834 0.037148270 0.009914587 0.009646656 0.006430774 0.061281831
#>  [7] 0.007204461 0.006662164 0.011014689 0.009283508 0.004427790 0.025347226
#> [13] 0.027171225 0.039443796 0.001586129 0.007667011 0.080059790

## the function mix.vcov() should be used for dataset